Moves ClientRequest utilities into their own directory
